TIP-1556: move feature flags in test namespace to be more epxlicit
