+++ src/Akeneo/Platform/Bundle/FeatureFlagBundle/Resources/config/test_behat_services.yml
@@ -1,7 +1,7 @@
 services:
     feature_flags:
         public: true
-        class: 'Akeneo\Platform\Bundle\FeatureFlagBundle\Internal\FilePersistedFeatureFlags'
+        class: 'Akeneo\Platform\Bundle\FeatureFlagBundle\Internal\Test\FilePersistedFeatureFlags'
         arguments:
             - '@akeneo.feature_flag.service.registry'
             - '%kernel.cache_dir%/'
