+++ src/Akeneo/Platform/Bundle/FeatureFlagBundle/Resources/config/test_fake_services.yml
@@ -1,6 +1,6 @@
 services:
     feature_flags:
         public: true
-        class: 'Akeneo\Platform\Bundle\FeatureFlagBundle\Internal\InMemoryFeatureFlags'
+        class: 'Akeneo\Platform\Bundle\FeatureFlagBundle\Internal\Test\InMemoryFeatureFlags'
         arguments:
             - '@akeneo.feature_flag.service.registry'
